Installer: streamline API.
To improve consistency and avoid confusion as different contributors update the installer, this patch enforces a very simple mental model: all commands run as-is on install-dir. Sending "yes" doesn't work for all commands, so each one should use the appropriate options for unattended installation. The cwd parameter is removed in favor of explicit 'cd' usage. Execution of bash string commands is also done explicitly. This have an added bonus of making the install logs cleaner, easier to read and easier to retype for testing.